For D-Bus de/marshalling types, use int instead of KGlobalAccel::MatchType
And cast to/from int in the two places where that is needed. Otherwise
the globalShortcutsByKey() method is silently skipped, and isn't visible on
the org.kde.KGlobalAccel D-Bus interface.

It turns out that to test changes made in kglobalaccel d-bus interface, you
need to kill the system /usr/bin/kglobalaccel process, and run the one from
the build-dir, otherwise the session d-bus won't pick up your changes.

To test:
- List all methods on the interface:
  `qdbus org.kde.kglobalaccel /kglobalaccel`
- The followin messages are printed in the system log/journal:
Skipped method "globalShortcutsByKey" : Unregistered input type in parameter list: KGlobalAccel::MatchType
Skipped method "globalShortcutsByKey" : Unregistered input type in parameter list: KGlobalAccel::MatchType

BUG: 454704
FIXED-IN: 5.95

+++ src/kglobalaccel.cpp
@@ -588,7 +588,7 @@ QList<KGlobalShortcutInfo> KGlobalAccel::getGlobalShortcutsByKey(const QKeySeque
 
 QList<KGlobalShortcutInfo> KGlobalAccel::globalShortcutsByKey(const QKeySequence &seq, MatchType type)
 {
-    return self()->d->iface()->globalShortcutsByKey(seq, type);
+    return self()->d->iface()->globalShortcutsByKey(seq, static_cast<int>(type));
 }
 
 bool KGlobalAccel::isGlobalShortcutAvailable(const QKeySequence &seq, const QString &comp)
